cmd/snapctl: move the sources around
