+++ .golangci.yml
@@ -100,7 +100,7 @@ linters:
         path: asserts/findwildcard_test.go
       - linters:
           - testpackage
-        path: cmd/snapctl/main_test.go
+        path: cmd/snapctl/tool/snapctl/main_test.go
       - linters:
           - testpackage
         path: cmd/snap-update-ns/sorting_test.go
