+++ cmd/snapctl/tool/snapctl/main_test.go
@@ -17,7 +17,7 @@
  *
  */
 
-package main
+package snapctl
 
 import (
 	"bytes"
